Intermedio 1

+++ FirmwareTeen/Definitions.h
@@ -75,9 +75,9 @@ const uint8_t NUMBERMISSEDCLOCKS = 3; /// Number of missed clock before stoppig 
 
 // Analog INputs resolution
 const uint8_t ANBITS = 16;				/// Bits read 8on ADC during AnalogRead
-const int32_t ANRANGEMAX = 4095;		/// Maximum range for Input. On Teensy out of the 16 bits read first 4 are discarded.
-const int32_t DEFMINDAC = 4095;	/// Default CV Minimum DAC
-const int32_t DEFRANGEDAC = -4095; /// Default CV Range DAC
+const int16_t ANRANGEMAX = 4095;		/// Maximum range for Input. On Teensy out of the 16 bits read first 4 are discarded.
+const int16_t DEFMINDAC = 4095;	/// Default CV Minimum DAC
+const int16_t DEFRANGEDAC = -4095; /// Default CV Range DAC
 
 #ifdef CVTHING
 const uint8_t PINENCA = 7;
